fix
- add `res.EpisodeWikiInfo` to `routes/schemas.ts`
- put validation functions under `lib/utils/`
